Fix: Clipboard and application frontend rewised.
