+++ ui/sleek-ui/widgets/input.slint
@@ -141,7 +141,7 @@ export component UInput inherits Rectangle {
         }
     ]
 
-	// Reexport the properties of the TextInput widget.
+    // Reexport the properties of the TextInput widget.
 	in-out property <brush> i-color <=> input.color;
     in-out property <string> i-font-family <=> input.font-family;
     in-out property <length> i-font-size <=> input.font-size;
@@ -171,6 +171,12 @@ export component UInput inherits Rectangle {
     }
     public function cut() {
         input.cut();
+    }
+    public function grab-focus() {
+        input.focus();
+    }
+    public function remove-focus() {
+        input.clear-focus();
     }
     public function paste() {
         input.paste();
